config/requirements/prod.txt
argon2_cffi==23.1.0
Django==4.2.7
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-prettyjson==0.4.1
django-stdimage==6.0.2
djangorestframework==3.14.0
djutils @ git+https://github.com/iagocanalejas/djutils.git@master
drf-spectacular==0.27.1
environs==10.3.0
gunicorn==21.2.0
inquirer==3.2.3
jsonschema==4.21.1
matplotlib==3.8.2
numpy==1.26.3
opencv-python==4.9.0.80
openpyxl==3.1.2
pandas==2.2.0
Pillow==10.2.0
psycopg2-binary==2.9.9
pytesseract==0.3.10
python-memcached==1.59
pyutils @ git+https://github.com/iagocanalejas/pyutils.git@master
PyYAML==6.0.1
requests==2.31.0
rscraping @ git+https://github.com/iagocanalejas/rscraping.git@master
setuptools==69.0.3
simplemma==0.9.1
